requirements.txt
asgiref==3.2.10
autopep8==1.5.3
dj-database-url==0.5.0
Django==3.1.1
django-bootstrap-modal-forms==2.0.0
django-ckeditor==5.9.0
django-crispy-forms==1.9.1
django-js-asset==1.2.2
django-markdown==0.8.4
django-popup-field==0.1.6
gunicorn==20.0.4
importlib-metadata==1.7.0
Markdown==3.2.2
Pillow==7.2.0
pycodestyle==2.6.0
python-decouple==3.3
pytz==2020.1
sqlparse==0.3.1
toml==0.10.1
Unipath==1.1
whitenoise==5.1.0
zipp==3.1.0
